Introduce datatype conversion API and large changes to the informix_fdw infrastructure.
This fairly large commit introduces an API for datatype conversion from Informix
to PostgreSQL types. This works for most of the basic datatypes at the moment,
support for BLOBs and advanced opaque datatypes is still missing.

Also, shift many of the declarative calls from ifxBeginForeignScan() to
ifxPlanForeignScan(), since only there we can use row and I/O estimates
returned by a declared cursor. This also involves introducing a serializable
infrastructure, so that parameters retrieved in ifxPlanForeignScan() can be
pushed down into the executor for ifxBeginForeignScan(). One side effect
of these code shifting is, that ifxPlanForeignScan() is now responsible
to prepare and declare the cursor (but not open it!), which was formerly
the task of ifxBeginForeignScan.