Make ifx_predicate_tree_walker() to not consider unsupported operands during examination.
Formerly, the code first added all predicate into a list for pushdown, filtering
out unsupported expressions afterwards. It turned out, that this is far to
complicated to parse them through later in ifxFilterQuals(), so filter them
out immediately by not even considering them for pushdown. This makes the code
in ifxFilterQuals() much easier to maintain and gives me a better feeling for
further improvements.