Further improvements on predicate pushdown feature.
Formerly the code didn't cope with foreign tables which are
part of a JOIN (e.g. <local rel> JOIN <foreign rel>). When retrieving
the original expression text via pg_get_expr(), we got varnos referencing the
RTE of the foreign table with the RT list of the whole query. The current
code handles this by rewriting the varnos within a copy of the expression
node, and pass its own deparse context to deparse_expression().

This seems to work with JOINs and subqueries after some tests, however,
further testing is required.