Repair BIGINT/INT8 datatype handling.
The original coding was under the impression, that INT8 and BIGINT
types are handled by ifx_int8_t, which was obviously wrong
(discovered during deeper testing of the current code). Fix the
conversion routines to map BIGINT properly. While at it,
make proper initialization to IfxConnectionInfo, which fails to
set query and tablename field values to a proper initialization value (NULL).