Further improvements for Informix cursor handling.
Originally, querying tables with BLOBs wasn't possible in case we
are using a SCROLL cursor. Additionally, self JOINs on the same
foreign informix tables weren't supported, since we didn't allocate
separate cursors for them.

Solve the problem with foreign tables having BLOBs by re-opening
a NO SCROLL cursor. This raises problems in informix database not
having logging enabled. Without proper transaction isolation level,
we might get dirty reads from a ReScan event, if the data has changed
on the foreign server in the meantime. Issue a warning in case someone
gets such a situation. With logging enabled, we use a serialized
transaction isolation level, which will guarantuee a consistent
cursor result in case we have a ReScan event. Since we reuse the
cursor definition for DESCRIBE during planning and execution, we
can't simply determine wether a SCROLL or NO SCROLL cursor is
suitable for the current foreign table. Introduce a new FDW variable
enable_blobs, which is required to be specified in case you want to
query foreign informix tables with BLOBs. If not specified, the FDW still
uses a SCROLL cursor per default, which will lead to an error in case
BLOB columns are involved.

The problem with self JOINs is solved by allocating a CURSOR private
to each *scan* of the foreign informix table, not the table itself
(which was the original coding). To accomplish this, generate a refid
from the current scan state and attach it to the identifier of the cursor
(in fact, the CURSOR itself isn't sufficient to get a unique name
this way, we also need to generate such identifier for DESCRIBE et al).