Introduce helper functions to terminate named informix connections.
ifx_fdw_close_connection() takes the informix connection identifier
(which could be taken from the output of ifx_fdw_get_connections())
and terminates it. This could be used to reset informix database
connections when the FDW is used in connection poolers for example.

While at it, tackle the long outstanding issue that we couldn't join
informix foreign tables living in different informix databases/servers.
This was because we didn't set the current connection for each
call in ifxIterateForeignScan().

This introduces a noticable overhead now and we need to overhaul
the ifxMakeConnectionInfo() infrastructure a little to make it faster.
I leave that for later...