Fix long-standing thinko in SQLDA data buffer initialization.
ifxGetColumnAttributes() is responsible to return a valid calculated
row size, suitable to be used to allocate a palloc'ed buffer, which
is assigned to the former created SQLDA structure. However, it failed
to calculate a proper value, since it missed to honour word aligned
calculation of offsets into that buffer. This resulted sometimes in
a too small row size, leading to overflowing the data buffer under some
circumstances (e.g, if you select smallint columns). Fix it by assigning
a proper aligned row size via rtypalign().