Initialize coninfo->username properly.
The former coding was unsafe and segfaulted in case someone
omits the username in the user mapping associated to the foreign
server.

We might consider to throw an error in case the associated
user mapping forgot to specify a correct username instead, however,
another option would be to just map the current role name to the
informix user we want to use for login. I'm not sure right now what
to do, so initialize the username to an empty string and let the code
throw an connection error since no valid username was supplied.