Filter out not-yet-supported datatypes from pushdown.
Currently temporal types are tricky, since we use the internal
string representation of a datatype value by calling the ruleutils
API. This only works for datatypes, which have a compatible text
representation in Informix. We need to improve this in the future, but
for now just filter them out to avoid pushing them down to the remote
Informix server. This looks much better than throwing an syntax error.