Teach predicate pushdown to correctly identify OR'ed sub-clauses.
Formerly the pushdown code didn't care enough about OR'ed subclauses
in predicates identified to be able to be pushed down to the remote server.
This could lead to wrong WHERE conditions, since omitting clauses
from a OR boolean expression is obviously wrong.

The current fix is to just return an empty filter from ifxFilterQuals()
in case a OR BoolExpr is removed from the predicates list. It's possible to
make this whole thing smarter, but for now we keep it that way to make
the filter conditions safe.