Implement full support for MONEY and NUMERIC type mapping.
This allows to specify foreign tables with columns having either
a mapping from MONEY or NUMERIC to both corresponding Informix types.

There's still the locale issue in connection of having a MONEY column
on the local and foreign table. We currently don't enforce any locale
compatibility between PostgreSQL and Informix when using a MONEY type
on both sides, so it's very likely to get confusing results when at least
the lc_monetary settings locally does not agree with the foreign Informix side.
Users should be very well aware about this issue, so mention it in
the README explicitely as a caveat.