Improve IMPORT FOREIGN SCHEMA support for temporal datatypes.
Commit 7ea33b1f539ef0d66ebfc5299a165307c7566c32 introduced the whole
infrastructure for IMPORT FOREIGN SCHEMA, but it's not yet complete
regarding datatype mapping. There are still corner cases around. This
commit tackles datatype mapping of Informix DATETIME and INTERVAL columns
to equivalent PostgreSQL types TIMESTAMP and INTERVAL. The current implementation
tries to honor the precision of an DATETIME or INTERVAL  when
mapping the foreign table locally. Since Informix has slight differences
there, not everything can be fully recognized. In this case the FDW tries
to choose just TIMESTAMP or INTERVAL without any qualifier.

This is just an attempt to cover as much as possible when matching datatypes
to foreign tables. It's more or less a feature to document as much as
possible with a foreign table declaration, so that an user get's an idea what the
foreign dataset looks like.