Fix calculation of required memory size when retrieving a VARCHAR column with maximum length.
ifxGetColumnAttributes() formerly wasn't defensive against rtypmsize() overflowing
when calculating the required memory size to fetch a VARCHAR(255) column value.
This led to "Character host variable is too short for the data." errors during
querying a foreign table with such columns. Fix that but just ignoring the formerly
returned size from rtypmsize() and reassign the maxlen value already got by the
DESCRIBE command within the SQLDA structure.

This behavior of rtypmsize() is not documented anywhere in the ESQL/C docs i can
find on the web, so i'm not sure wether this is just a bug in the CSDK or an
oversight by me.