Don't leak password string into log messages.
When checking for duplicates in options specified to the FDW,
we accidently leaked the password string into the log message
in case the 'password' options was specified twice.

While at it, convert the ERRCODE_SYNTAX_ERROR into
ERRCODE_FDW_INVALID_OPTION, which looks more appropiate.
This was already true for most of the checks in
ifxGetOptionDups(), but some of them were still using the former.