Fix a long standing stupid error by using uninitialized password strings.
This could lead to backend crashes when leaving out the password option
to the informix database connection. To fix this, initialize the password
string to an empty string like we already do for the username. I don't know
why this wasn't reported before, but maybe since password is a required option
to login no one ever had this issue before.

Per report from Jeff Ross (rossj-cargotel).