Teach IMPORT FOREIGN SCHEMA to set proper foreign table options.
When importing a foreign table definition with BYTES or TEXT
column types, we forgot to set the enable_blobs option.
To teach the Informix FDW to do it correctly,
ifxGetForeignTableDetails() needs to get the IfxImportTableDef
structure specified.

While at it, remove the useless ImportForeignSchemaStmt
argument from it, it wasn't used anymore.