Better pushdown handling for constant values.
Issue https://github.com/credativ/informix_fdw/issues/11 brought up
cases with certain datatypes like bpchar which were ill-conceived
earlier.

The idea to blindly pass predicates down to PostgreSQL's own deparse machinery
is not fully functional for all types: e.g. bpchar is a datatype which
will be decorated with a cast expression to force it away from
being text and to match its original column type. Further, Florian
Apolloner griped about not correctly handling the typmod of types
like varchar. Unsurprisingly, Informix is not very happy about this
kind of SQL expressions.

This is a PoC, which makes the handling for Const * nodes in the
pushdown machinery explicit. It can simply improved for other
expressions (such as compatible function expression et al.), but
its primary purpose is to test Florian's and rossj-cargotel's use cases
and that it doesn't break any things.