Support REL_11_STABLE.
get_relid_attribute_name() is gone and is replaced by
get_attname().

ExplainPropertyFloat() requires a new "unit" argument.

Move old-style tuple descriptor usage and access to its attributes
to the "new" TupleDescAttr() macro. We are encouraged to do so since a while.

Since TupleDescAttr() is backpatched to 9.3 and appears first in 9.3.19,
we need to make sure we use the correct definition depending on the
PostgreSQL version we build against.

This makes it important to use a correct build of the informix_fdw.so
library against 9.3 instances.