Adjust regression tests.
Starting with PostgreSQL 12 has changed the way we handle float digits, so
we are forced to reflect these changes in the regression tests of the
FDW (e.g. extra_float_digits defaults to 1 now).

To avoid false differences between the major versions, we are going to
maintain two alternative expected files, the regression test suite then picks
the one with the closest matches.