Support for PostgreSQL 14.
AddForeignUpdateTargets() has changed its signature. This comes along with a
new approach on registering resjunk columns to the target list (in this case
the encoded ROWID of the foreign tuple). The new approach is to use
add_row_identity_var(). We put that in an additional #ifdef'ed block and i'm
under the impression it's time to get rid of some of the older stuff #ifdef'ed,
too...but let's hold that back for another commit.

While at it, get rid of the (long ago) deprecated usage of bzero() and
switch to memset().