Add build and test framework.
I was tired of having all kind PG installations around to test and
build the FDW, so this automates the whole building and test environments with `podman` and containers.

Especially the PG containers are fully equipped with all required development files and can
easily be used to debug the Informix FDW extension. There is a `build.sh` script included, which setups PostgreSQL, compiles the FDW and runs the regression tests (currently `informix_fdw_tx` transactional tests only).

The Informix database runs within a customized image from

https://hub.docker.com/r/ibmcom/informix-developer-database

See the included README for more detailed instructions on how to use it.

I haven't tested it with docker so far.