Further cleanup of commit 098aae63788e6c6d576f3006133f9a5c31a9259f.
The missing break out masked another bug we had in this area: We cannot just
pass the string value retrieved from the locator in `convertIfxSimpleLO()` to
the type input function without copying it properly into a null terminated character
buffer, otherwise we scribble on memory again.

This wasn't visible before, since the missing break just handed this over to
the BYTEA conversion, which did the right thing by copying the buffer into a
properly allocated buffer.

Fix this mess by handling this with a proper character buffer and take care
its properly zero-initialized.