Fix type mapping for BIGSERIAL columns.
Originally we mapped BIGSERIAL (or IFX_BIGSERIAL) SQL types from Informix
to ESQL/C's CINT8TYPE, which turns out to be wrong. Issue #1 revealed that this
causes undefined behaviour, since this doesn't initialize the targeted
`ifx_int8_t` host variables correctly. To be honest, it was by accident that
this was originally working, since we had regression tests for this. By having
BIGSERIAL mapped to `ifx_int8_t`, we never get a properly initialized `sign`
field of the struct. This was never a problem when we had any positive value here
(everything >= 1 seems to be considered as a positive number), but once we end up
with a 0 here, every transform function from ESQL/C takes this as an undefined
number.

The results are as described in issue #1. The fix is easy, just map BIGSERIAL
correctly to CBIGINTTYPE, making sure it calls the correct conversion routines.

Adjust the regression tests for BIGSERIAL to have a test with a negative value.