Fix github issue #4.
This is a fix for a longstanding oversight in ifx_predicate_tree_walker():

When examining baserestrictinfo expression nodes for push down, we forgot
to make sure to preserve the precedence of OR'ed BoolExpr operators. E.g.
when we have this:

col1 = <val> AND (col2 = <val> OR col2 = <val>)

we silently decode the BoolExpr without the round brackets, getting

col1 = <val> AND col2 = <val> OR col2 = <val>

which is obviously wrong.

To fix this, we teach ifx_predicate_tree_walker() to push IfxPushdownOprInfo
with corresponding context information into the pushdown context, so that
in the decoding phase in ifxFilterQuals() corresponding round brackets are
produced.